+++ lib/jsonapi_rails/params_to_object.rb
@@ -8,15 +8,15 @@ module JsonApiRails
     attr_reader :attributes
     # Array of attribute names not set on the Resource but that should be allowed
     # Useful for transient attributes, such as a credit card number
-    attr_reader :whitelisted
+    attr_reader :permitted
 
-    def initialize(data_hash, ar_relation = nil, resource_class = nil, whitelisted = [])
+    def initialize(data_hash, ar_relation = nil, resource_class = nil, permitted = [])
       validate_json(data_hash)
 
       data = data_hash[:data]
       @object = setup_object(data, ar_relation)
       @resource_class = resource_class
-      @whitelisted = whitelisted
+      @permitted = permitted
 
       @attributes = Hash(data[:attributes])
       @relationships = Hash(data[:relationships])
@@ -95,7 +95,7 @@ module JsonApiRails
     #   the attribute
     def assign_attributes
       attributes.each_with_object({}) do |(attr, value), hsh|
-        unless whitelisted?(attr)
+        unless permitted?(attr)
           message = "`#{resource.class}' does not have attribute " \
                     "`#{attr.to_s.gsub('=', '')}'"
           fail UnknownAttributeError.new(message)
@@ -255,13 +255,13 @@ module JsonApiRails
 
     # Inspects the model using `attribute_names` and selects the attributes
     # which are also present in the Resource's `fields_array`. Alternatively,
-    # properties from the model can be whitelisted, and will also be returned.
+    # properties from the model can be permitted, and will also be returned.
     #
     # @return [Array<String>] list of attributes available to assign to the
     # underlying model
     def assignable_attribute_names
-      stored_attributes = object.attribute_names.select &method(:whitelisted?)
-      transient_attributes = whitelisted.select do |attribute_name|
+      stored_attributes = object.attribute_names.select &method(:permitted?)
+      transient_attributes = permitted.select do |attribute_name|
         check_method("#{attribute_name}=")
         true
       end.map(&:to_s)
@@ -285,10 +285,10 @@ module JsonApiRails
     end
 
     # Checks for the attribute's presence in the Resource's fields array or the
-    # whitelist
-    def whitelisted?(attribute_name)
+    # permitted whitelist
+    def permitted?(attribute_name)
       resource.fields_array.include?(attribute_name.to_sym) ||
-      whitelisted.include?(attribute_name.to_sym)
+      permitted.include?(attribute_name.to_sym)
     end
   end
 end