+++ lib/jsonapi_rails/parser.rb
@@ -3,11 +3,11 @@ module JsonApiRails
   # `ParamsToObject`, then extracting the model and passing it to the given
   # block.
   class Parser
-    def initialize(params_hsh, ar_relation = nil, resource_class = nil, whitelisted = [])
+    def initialize(params_hsh, ar_relation = nil, resource_class = nil, permitted = [])
       @wrapper = ParamsToObject.new params_hsh,
                                     ar_relation,
                                     resource_class,
-                                    whitelisted
+                                    permitted
     end
 
     # Extracts the model from the wrapper, then passes it to the given block.