+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+v 0.4.0-beta
+Renames the `whitelisted` option to `permitted`
+
 v 0.3.2-beta
-Adds the `whitelisted` option to `#parse_json_api_params` 
+Adds the `whitelisted` option to `#parse_json_api_params`
 
 v 0.3.1-beta
 Version bump for json api ruby dependency. Fixes duplication in the